requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
face-recognition==1.3.0
opencv-python==4.5.3.56
numpy==1.21.2
mysql-connector-python==8.0.26
python-dotenv==0.19.0
flask-login==0.5.0
flask-wtf==0.15.1
werkzeug==2.0.1
pillow==8.3.2
scikit-learn==0.24.2
pandas==1.3.3
openpyxl==3.0.9 